+++ databricks/sdk/__init__.py
@@ -6,7 +6,7 @@ from typing import Optional
 import databricks.sdk.core as client
 import databricks.sdk.dbutils as dbutils
 import databricks.sdk.service as service
-from databricks.sdk import azure
+from databricks.sdk.databricks import azure
 from databricks.sdk.credentials_provider import CredentialsStrategy
 from databricks.sdk.data_plane import DataPlaneTokenSource
 from databricks.sdk.mixins.compute import ClustersExt
