+++ databricks/sdk/databricks/core.py
@@ -3,7 +3,7 @@ from typing import BinaryIO
 from urllib.parse import urlencode
 
 from ._base_client import _BaseClient
-from .config import *
+from .databricks.config import *
 # To preserve backwards compatibility (as these definitions were previously in this module)
 from .credentials_provider import *
 from .errors import DatabricksError, _ErrorCustomizer
