+++ databricks/sdk/databricks/credentials_provider.py
@@ -20,7 +20,7 @@ from google.auth import impersonated_credentials  # type: ignore
 from google.auth.transport.requests import Request  # type: ignore
 from google.oauth2 import service_account  # type: ignore
 
-from .azure import add_sp_management_token, add_workspace_id_header
+from .databricks.azure import add_sp_management_token, add_workspace_id_header
 from .oauth import (ClientCredentials, OAuthClient, Refreshable, Token,
                     TokenCache, TokenSource)
