+++ tests/testdata/test_casing.py
@@ -1,6 +1,6 @@
 import pytest
 
-from databricks.sdk.casing import Casing
+from databricks.sdk.databricks.casing import Casing
 
 
 @pytest.mark.parametrize(
